// frontend/src/app/api/model/flight.ts
/**
 * FlightBookingAPI
 * API aplikacije za kupovinu avionskih karata
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Departure } from './departure';
import { SoldTicket } from './soldTicket';
import { Arrival } from './arrival';


/**
 * Podaci o letu.
 */
export interface Flight { 
    /**
     * Identifikator leta.
     */
    flightId: string;
    /**
     * Podaci o kupljenim kartama.
     */
    soldTickets: Array<SoldTicket>;
    /**
     * Broj ukupnih karata na prodaju.
     */
    totalTickets: number;
    /**
     * Broj preostalih karata za prodaju.
     */
    availableTickets: number;
    /**
     * Trenutna cena jedne karte na prodaju.
     */
    ticketPrice: number;
    /**
     * Ukupna vrednost svih prodatih karata.
     */
    totalPrice: number;
    departure: Departure;
    arrival: Arrival;
    /**
     * Oznacava da li je proslo vreme polaska leta.
     */
    passed: boolean;
    /**
     * Podatak o tome da li je let otkazan.
     */
    canceled: boolean;
}